+++ src/main/java/com/bsuir/test/service/book/BookService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface BookService {
     List<Book> getAllBooks();
     void saveBook(Book book);
-    Book getBookById(Long id);
-    void deleteBookById(Long id);
+    Book getBookById(int id);
+    void deleteBookById(int id);
     Page<Book> findPaginated(int pageNo, int pageSize, String sortField, String sortDirection);
 }
